Dev: unittests: Add dedicated unit tests for storage_utils
- Add test/unittests/test_storage_utils.py with tests for all
  functions in crmsh/storage_utils.py
- Move storage_utils test cases from test_utils.py to the new file
- Add new tests for previously untested functions:
  - get_dev_uuid_by_blkid (UUID present and absent)
  - get_dev_uuid fallback path via blkid
  - detect_duplicate_device_path (no duplicate and duplicate cases)
  - is_dlm_running (default args and on_node arg)
- Remove storage_utils import from test_utils.py (no longer needed)
