+++ data-manifest
@@ -215,6 +215,7 @@ test/unittests/test_sbd.py
 test/unittests/test_scripts.py
 test/unittests/test_service_manager.py
 test/unittests/test_sh.py
+test/unittests/test_storage_utils.py
 test/unittests/test_time.py
 test/unittests/test_ui_cluster.py
 test/unittests/test_ui_corosync.py
